% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compute_count}
\alias{compute_count}
\title{Count unique observations.}
\usage{
compute_count(x, x_var, w_var = NULL)
}
\arguments{
\item{x}{Dataset-like object to count. Built-in methods for data frames,
grouped data frames and ggvis visualisations.}

\item{x_var,w_var}{Names of x and weight variables.}
}
\description{
Count unique observations.
}
\examples{
mtcars \%>\% compute_count(~cyl)

# Weight the counts by car weight value
mtcars \%>\% compute_count(~cyl, ~wt)

# If there's one weight value at each x, it effectively just renames columns.
pressure \%>\% compute_count(~temperature, ~pressure)
}
\seealso{
\code{\link{compute_bin}} For counting cases within ranges of
  a continuous variable.
}

