% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R, R/compute.R
\name{compute_stat.bin_fixed}
\alias{compute_stat}
\alias{compute_stat.bin_breaks}
\alias{compute_stat.bin_fixed}
\title{Compute a statistical transformation}
\usage{
\method{compute_stat}{bin_fixed}(params, x, ..., w = NULL)

\method{compute_stat}{bin_breaks}(params, x, ..., w = NULL)

compute_stat(params, x, ..., w = NULL)
}
\arguments{
\item{params}{A parameter object generated by of the other functions in
this packagea}

\item{x}{X position. This is required by all stats. \code{y}, \code{z},
and \code{w} should always be the same length as \code{x}.}

\item{...}{Additional arguments passed on to stat. Many stats take
\code{y} and \code{z} values.}

\item{w}{Weights. Optional vector of weights.}
}
\description{
Compute a statistical transformation
}

