% Generated by roxygen2 (4.0.2): do not edit by hand
\name{restore}
\alias{restore}
\title{Restore S3 class properties.}
\usage{
restore(old, new)
}
\arguments{
\item{old}{Old S3 vector}

\item{new}{New S3 vector. Should be same \code{\link{typeof}()} as
\code{old}.}
}
\description{
When an S3 class is built on top of an atomic vector, it's often resonable
to do some computation on the underlying atomic vector. This generic allows
you to restore important attributes from the original vector to the new
vector, so that it continues to behave as expected.
}
\examples{
restore(Sys.Date(), 1)
restore(Sys.time(), 1)

dt <- difftime(Sys.time(), Sys.time() - 100)
restore(dt, 10)
}

