% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compute_quantile_vec}
\alias{compute_quantile_vec}
\title{Compute quantiles (vectorised).}
\usage{
compute_quantile_vec(x, w = NULL, probs = seq(0, 1, 0.25))
}
\description{
Compute quantiles (vectorised).
}
\examples{
compute_quantile_vec(1:100)

# Specify weights with second argument
compute_quantile_vec(1:100, 100:1)
}
\keyword{internal}

