% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-2d.R, R/bin.R, R/compute.R
\name{compute_stat.bin_2d_fixed}
\alias{compute_stat}
\alias{compute_stat.bin_2d_fixed}
\alias{compute_stat.bin_breaks}
\alias{compute_stat.bin_fixed}
\title{Compute a statistical transformation}
\usage{
\method{compute_stat}{bin_2d_fixed}(params, x, y, ..., w = NULL)

\method{compute_stat}{bin_fixed}(params, x, ..., w = NULL)

\method{compute_stat}{bin_breaks}(params, x, ..., w = NULL)

compute_stat(params, x, ..., w = NULL)
}
\arguments{
\item{params}{A parameter object generated by of the other functions in
this packagea}

\item{x}{X position. This is required by all stats.}

\item{y}{Y position. This is required by all 2d stats.}

\item{...}{Additional arguments passed on to stat.}

\item{w}{Optional weights. If supplied, must be a numeric vector the
same length as \code{x}.}
}
\description{
Compute a statistical transformation
}

