% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_stat}
\alias{compute_stat}
\title{Compute a statistical transformation}
\usage{
compute_stat(params, x, y, z, w = NULL)
}
\arguments{
\item{params}{A parameter object generated by of the other functions in
this packagea}

\item{x}{X position. This is required by all stats. \code{y}, \code{z},
and \code{w} should always be the same length as \code{x}.}

\item{y}{Y position. This is required by all 2d stats.}

\item{z}{Z value. This is required by all stats that compute a summary.}

\item{w}{Weights. Optional vector of weights.}
}
\description{
Compute a statistical transformation
}

